requirements.txt
python-telegram-bot>=22,<23
python-dotenv
httpx>=0.27,<1
